Pre-sort QueryIter layers
`sort_layers` assumes that the layers are already partially sorted so
we must pre-sort the layers. This is a hack identical to the "HAXXX"
code in `HighlightIter`.

Ideally we would be able to say

    layers.sort_unstable_by_key(|layer| layer.sort_key());

However `Vec::sort_unstable_by_key` only gives immutable borrows of the
elements (for correctness) and `IterLayer::sort_key` needs a mutable
borrow to peek captures.

So for now we approximate the sort order using the first range in the
layer and the depth descending. This doesn't work robustly for combined
injections however.